edited data types, removed unnecesary comments

+++ include/ob/Orderbook.hpp
@@ -6,6 +6,7 @@ using namespace std;
 #include "Trade.hpp"
 #include <vector>
 #include <algorithm>
+#include <stdint.h>
 
 namespace ob {
     class OrderBook {
@@ -17,16 +18,10 @@ namespace ob {
             // return 0 on success, -1 if order_id not found
             int cancel_order(int order_id); // -> Implement later
 
-            // adds a market order to OB -> executes trades immediately
-            // returns 0 on success, -1 on failure
-            // std::vector<Trade> add_market(const Order& order); // -> Implement later
-            // need a way to represent market orders
-
-
         private:
             // helper functions and other stuff for order matching and trade execution
-            int cur_trade_id = 0;
-            int cur_seq = 0;
+            uint64_t cur_trade_id = 0;
+            uint64_t cur_seq = 0;
 
             // for now use a vector to store orders, will need a more efficient structure for real implementation
             vector<Order> bids;
