+++ include/ob/Trade.hpp
@@ -1,11 +1,12 @@
 #pragma once
+#include <stdint.h>
 
 struct Trade{
-    int trade_id;
+    uint64_t trade_id;
     int price_tick;
-    int quantity;
-    int taker_order_id; // id of the order that initiated the trade
-    int maker_order_id; // id of the order that was resting in the book
+    uint16_t quantity;
+    uint64_t taker_order_id; // id of the order that initiated the trade
+    uint64_t maker_order_id; // id of the order that was resting in the book
 };
 
 // Trades (trade_id, price, qty, taker_order_id, maker_order_id)
